backend/requirements.txt
# Sofia V2 Realtime DataHub - Production Dependencies

# Core Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# HTTP Client & Async
httpx>=0.25.0
aiofiles>=23.2.1
asyncio-mqtt>=0.13.0

# Data Processing
orjson>=3.9.0
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0

# Configuration & Validation
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Database (Optional TimescaleDB)
asyncpg>=0.28.0
sqlalchemy[asyncio]>=2.0.0

# WebSocket Clients
websocket-client>=1.6.0
python-socketio>=5.10.0

# Monitoring & Metrics
prometheus-client>=0.18.0
structlog>=23.2.0

# Date/Time handling
python-dateutil>=2.8.0
pytz>=2023.3

# RSS/XML Processing
feedparser>=6.0.10
lxml>=4.9.0

# Crypto/Security
cryptography>=41.0.0

# Utilities
backoff>=2.2.0
tenacity>=8.2.0
schedule>=1.2.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
mypy>=1.6.0